requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
pypdf
sentence-transformers